Create encoding and decoding text
Here is a python program that simply takes a message and a shift number as an inputs and then encodes the message by shifting each letter forward alphabetically as per the shift number. Again it decodes a message inputted by shifting each letter in the message backwards as per the shift number given.